// src/generated/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  superadmin: 'superadmin',
  bendahara: 'bendahara',
  admin_kegiatan: 'admin_kegiatan',
  admin_inventaris: 'admin_inventaris'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const TransactionType = {
  income: 'income',
  expense: 'expense'
} as const

export type TransactionType = (typeof TransactionType)[keyof typeof TransactionType]


export const ZisCategory = {
  zakat: 'zakat',
  infaq: 'infaq',
  shadaqah: 'shadaqah'
} as const

export type ZisCategory = (typeof ZisCategory)[keyof typeof ZisCategory]


export const DonationStatus = {
  pending: 'pending',
  verified: 'verified',
  rejected: 'rejected'
} as const

export type DonationStatus = (typeof DonationStatus)[keyof typeof DonationStatus]


export const EventStatus = {
  upcoming: 'upcoming',
  ongoing: 'ongoing',
  completed: 'completed',
  cancelled: 'cancelled'
} as const

export type EventStatus = (typeof EventStatus)[keyof typeof EventStatus]


export const InventoryCondition = {
  baik: 'baik',
  rusak_ringan: 'rusak_ringan',
  rusak_berat: 'rusak_berat',
  hilang: 'hilang'
} as const

export type InventoryCondition = (typeof InventoryCondition)[keyof typeof InventoryCondition]


export const LoanStatus = {
  borrowed: 'borrowed',
  returned: 'returned',
  overdue: 'overdue'
} as const

export type LoanStatus = (typeof LoanStatus)[keyof typeof LoanStatus]


export const Gender = {
  male: 'male',
  female: 'female'
} as const

export type Gender = (typeof Gender)[keyof typeof Gender]


export const MustahikCategory = {
  fakir: 'fakir',
  miskin: 'miskin',
  amil: 'amil',
  muallaf: 'muallaf',
  riqab: 'riqab',
  gharim: 'gharim',
  fisabilillah: 'fisabilillah',
  ibnu_sabil: 'ibnu_sabil'
} as const

export type MustahikCategory = (typeof MustahikCategory)[keyof typeof MustahikCategory]